# frontend/openapi.json
{
	"components": {
		"schemas": {
...
			"BillingStatusResponse": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/BillingStatusResponse.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"cancel_at_period_end": {
						"description": "Whether the subscription cancels at period end",
						"type": "boolean"
					},
					"current_period_end": {
						"description": "Current billing period end",
						"type": "string"
					},
					"limits": {
						"additionalProperties": {
							"format": "int64",
							"type": "integer"
						},
						"description": "Entitlement limits from the local subscription snapshot",
						"type": "object"
					},
					"period_start": {
						"description": "UTC month start for the usage counters",
						"type": "string"
					},
					"plan_id": {
						"description": "Plan ID",
						"type": "string"
					},
					"provider": {
						"description": "Billing provider",
						"type": "string"
					},
					"status": {
						"description": "Subscription status",
						"type": "string"
					},
					"usage": {
						"additionalProperties": {
							"format": "int64",
							"type": "integer"
						},
						"description": "Current-month usage counters",
						"type": "object"
					},
					"workspace_id": {
						"description": "Workspace ID",
						"type": "string"
					}
				},
				"required": [
					"workspace_id",
					"status",
					"cancel_at_period_end",
					"limits",
					"usage",
					"period_start"
				],
				"type": "object"
			},
			"BillingURLResponse": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/BillingURLResponse.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"id": {
						"description": "Provider object ID",
						"type": "string"
					},
					"url": {
						"description": "Redirect URL",
						"type": "string"
					}
				},
				"required": ["url"],
				"type": "object"
			},
...
			"CreateBillingCheckoutInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/CreateBillingCheckoutInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"plan_id": {
						"description": "Plan ID: starter, creator, or pro",
						"type": "string"
					},
					"workspace_id": {
						"description": "Workspace ID",
						"type": "string"
					}
				},
				"required": ["workspace_id", "plan_id"],
				"type": "object"
			},
			"CreateBillingPortalInputBody": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/CreateBillingPortalInputBody.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"workspace_id": {
						"description": "Workspace ID",
						"type": "string"
					}
				},
				"required": ["workspace_id"],
				"type": "object"
			},
...
			"ErrorDetail": {
				"additionalProperties": false,
				"properties": {
					"location": {
						"description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
						"type": "string"
					},
					"message": {
						"description": "Error message text",
						"type": "string"
					},
					"value": {
						"description": "The value at the given location"
					}
				},
				"type": "object"
			},
			"ErrorModel": {
				"additionalProperties": false,
				"properties": {
					"$schema": {
						"description": "A URL to the JSON Schema for this object.",
						"examples": ["https://example.com/schemas/ErrorModel.json"],
						"format": "uri",
						"readOnly": true,
						"type": "string"
					},
					"detail": {
						"description": "A human-readable explanation specific to this occurrence of the problem.",
						"examples": ["Property foo is required but is missing."],
						"type": "string"
					},
					"errors": {
						"description": "Optional list of individual error details",
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": ["array", "null"]
					},
					"instance": {
						"description": "A URI reference that identifies the specific occurrence of the problem.",
						"examples": ["https://example.com/error-log/abc123"],
						"format": "uri",
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"examples": [400],
						"format": "int64",
						"type": "integer"
					},
					"title": {
						"description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
						"examples": ["Bad Request"],
						"type": "string"
					},
					"type": {
						"default": "about:blank",
						"description": "A URI reference to human-readable documentation for the error.",
						"examples": ["https://example.com/errors/example"],
						"format": "uri",
						"type": "string"
					}
				},
				"type": "object"
			},
...
		}
	},
	"info": {
		"title": "OpenPost API",
		"version": "1.0.0"
	},
	"openapi": "3.1.0",
	"paths": {
...
		"/billing/checkout": {
			"post": {
				"operationId": "create-billing-checkout",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBillingCheckoutInputBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BillingURLResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Bad Request"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Forbidden"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					},
					"503": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Service Unavailable"
					}
				},
				"summary": "Create billing checkout",
				"tags": ["Billing"]
			}
		},
		"/billing/portal": {
			"post": {
				"operationId": "create-billing-portal-session",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBillingPortalInputBody"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BillingURLResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Bad Request"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Forbidden"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					},
					"503": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Service Unavailable"
					}
				},
				"summary": "Create billing portal session",
				"tags": ["Billing"]
			}
		},
		"/billing/status": {
			"get": {
				"operationId": "get-billing-status",
				"parameters": [
					{
						"description": "Workspace ID",
						"explode": false,
						"in": "query",
						"name": "workspace_id",
						"schema": {
							"description": "Workspace ID",
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BillingStatusResponse"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Bad Request"
					},
					"403": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Forbidden"
					},
					"422": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Unprocessable Entity"
					},
					"500": {
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorModel"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"summary": "Get billing status",
				"tags": ["Billing"]
			}
		},
...
	}
}